Create SimpleStorage.sol
This is the first contract made in the Cyfrin course. I added the comments and kept them there for carnality details. Maybe I will adjust those later with more expansion so it is readable for absolute beginners. 
The essence is a contract that takes from input for a name and number that the user can give in and then store it in the struct of Persons. Those are visible with mappings to see which person has which favorite number, a way to store a favoritenumber and how to retrieve it.